requirements.txt
pyyaml
colorama
jinja2
pytest
